header and footer fixes
