+++ src/components/Printable/index.tsx
@@ -1,6 +1,14 @@
 import { Close } from '@mui/icons-material'
-import { Box, Button, IconButton, Modal, type ModalProps } from '@mui/material'
+import {
+  Box,
+  Button,
+  IconButton,
+  Modal,
+  Typography,
+  type ModalProps
+} from '@mui/material'
 import FlexBox from '../FlexBox'
+import { useConfigData } from '../../utils/useConfigData'
 
 interface Props {
   children?: React.ReactElement | React.ReactElement[]
@@ -32,17 +40,43 @@ const style = {
   }
 }
 
-const buttonSx = {
-  position: 'fixed',
-  right: '10px',
-  top: '10px',
-  zIndex: '9'
+const Header = ({
+  configData
+}: {
+  configData: ConfigData | undefined
+}): React.ReactElement => {
+  return (
+    <Box textAlign='center' marginBottom={2}>
+      <Typography variant='caption'>{configData?.headerMarking}</Typography>
+    </Box>
+  )
+}
+
+const Footer = ({
+  configData
+}: {
+  configData: ConfigData | undefined
+}): React.ReactElement => {
+  return (
+    <Box
+      position='fixed'
+      left={0}
+      right={0}
+      bottom={0}
+      textAlign='center'
+      width='100%'
+      bgcolor='white'
+      zIndex={1000}>
+      <Typography variant='caption'>{configData?.headerMarking}</Typography>
+    </Box>
+  )
 }
 
 export default function Printable(
   props: Props & ModalProps
 ): React.ReactElement {
   const { children, onPrint, ...rest } = props
+  const configData = useConfigData()
 
   const handleClose = (): void => {
     rest?.onClose?.({}, 'escapeKeyDown')
@@ -57,11 +91,20 @@ export default function Printable(
 
   return (
     <Modal sx={style} hideBackdrop className='printable' {...rest}>
-      <Box width='100%' height='100%' overflow='auto'>
-        <IconButton sx={buttonSx} onClick={handleClose}>
-          <Close />
-        </IconButton>
-        <Box sx={{ padding: '30px' }}>{children}</Box>
+      <Box width='100%' height='100%' display='flex' flexDirection='column'>
+        <Box
+          display='flex'
+          justifyContent='center'
+          marginBottom='10px'
+          position='relative'>
+          <Header configData={configData} />
+          <Box position='absolute' top='10px' right='10px'>
+            <IconButton onClick={handleClose}>
+              <Close />
+            </IconButton>
+          </Box>
+        </Box>
+        <Box sx={{ padding: '30px', overflowY: 'auto' }}>{children}</Box>
         <FlexBox
           className='noprint'
           marginBottom='20px'
@@ -73,6 +116,7 @@ export default function Printable(
             Cancel
           </Button>
         </FlexBox>
+        <Footer configData={configData} />
       </Box>
     </Modal>
   )
